src/mpflash/pyproject.toml
[tool.poetry]
name = "mpflash"
version = "0.7.7"
description = "Flash and download tool for MicroPython firmwares"
authors = ["Jos Verlinde <jos_verlinde@hotmail.com>"]
license = "MIT"
readme = "README.md"
keywords = ["MicroPython", "firmware", "flash", "download", "UF2", "esptool"]
homepage = "https://github.com/Josverl/micropython-stubber/blob/main/src/mpflash/README.md"
repository = "https://github.com/Josverl/micropython-stubber"
classifiers = [
    "Programming Language :: Python :: Implementation :: MicroPython",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Software Development :: Build Tools",
]

[tool.poetry.scripts]
mpflash = "mpflash.cli_main:mpflash"

[tool.poetry.dependencies]
python =  ">=3.8.1,<4.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
loguru = "^0.7.2"
esptool = "^4.7.0"
jsonlines = "^4.0.0"
bincopy = "^20.0.0"
# strip-ansi = "^0.1.1"
rich-click = "^1.7.3"
psutil = "^5.9.8"
blkinfo = "^0.2.0"
pygithub = "^2.1.1"
platformdirs = "^4.2.0"
pyusb = "^1.2.1"
packaging = "23.2"
tenacity = "8.2.3"
mpremote = "^1.22.0"
inquirer = "^3.2.4"
libusb = {version = "^1.0.27", platform = "win32"}
jsons = "^1.6.3"
sentry-sdk = "^2.1.1"

[tool.poetry.group.test]
optional = true
[tool.poetry.group.test.dependencies]
coverage = ">=6.4.3,<8.0.0"
pytest = "^7.1.2"
pytest-github-actions-annotate-failures = ">=0.1.7,<0.3.0"
pytest-json-report = "^1.5.0"
pytest-metadata = ">=2.0.2,<4.0.0"
pytest-mock = "^3.10.0"
mock = "^4.0.3"
distro = "^1.8.0"
fasteners = "^0.19"
jsons = "^1.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
